+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 0.7.6 - 2019-XX-XX
+
+- Add variability in the instance context refresh interval to avoid 'burst' on the Gateway API
+- Change the default refresh interval from 2 minutes to 10 minutes
+
 # 0.7.5 - 2019-03-08
 
 - Fix Handlebars typescript declaration conflicts